상품 저장 로직 구현
user관련 태호가 머지해주기 전에는 임시로 만들어서 쓸거임. 추후 머지 되면 없어짐.

+++ src/main/java/com/gulbi/Backend/domain/rental/product/controller/CategoryController.java
@@ -29,7 +29,7 @@ public class CategoryController {
     }
 
     @GetMapping("/mcategory/{categoryId}")
-    public ResponseEntity<RestApiResponse> getMidCategory(@PathVariable("categoryId") Long categoryId){
+    public ResponseEntity<RestApiResponse> getMidCategory(@PathVariable("categoryId") Integer categoryId){
         List<CategoryProjection> list = categoryService.getBelowCategoriesByParentId(categoryId);
 
         RestApiResponse response = new RestApiResponse(CategorySuccessCode.GET_CATEGORY_SUCCESS,list);
